third_party/blink/web_tests/platform/generic/fast/forms/indeterminate-input-types-expected.txt

Test for Bug: 76995 - Checks whether indeterminate property works with various input types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radioInput.indeterminate is true
PASS checkboxInput.indeterminate is true
PASS dateInput.indeterminate is true
PASS emailInput.indeterminate is true
PASS fileInput.indeterminate is true
PASS numberInput.indeterminate is true
PASS passwordInput.indeterminate is true
PASS rangeInput.indeterminate is true
PASS searchInput.indeterminate is true
PASS submitInput.indeterminate is true
PASS telInput.indeterminate is true
PASS timeInput.indeterminate is true
PASS urlInput.indeterminate is true
PASS weekInput.indeterminate is true
PASS successfullyParsed is true

TEST COMPLETE
